Refactor skeleton components and update import paths across features
